refactor: Results 클래스 생성
