+++ src/test/java/laddergame/domain/user/UsersTest.java
@@ -14,10 +14,4 @@ public class UsersTest {
                 .asList()
                 .hasSize(4);
     }
-
-    @DisplayName("실행 결과의 개수가 참여 인원수와 맞지 않으면 에러 발생")
-    @Test
-    void no_match_user_number_create_results() {
-        assertThatIllegalArgumentException().isThrownBy(() -> new Users("pobi,honux,crong,jk", "꽝,30000"));
-    }
 }
